resolve_components.py completed
